refactor: separate out and improve core game impl
